backend/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
flask-migrate
python-dotenv
